+++ packages/plugin-app/src/actions/add.ts
@@ -1,0 +1,4 @@
+/**
+ * 添加插件
+ * @author: sunkeysun
+ */