+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @inventorjs/cli
 
+## 1.5.4
+
+### Patch Changes
+
+- 82aaad2: support context import plugin
+- Updated dependencies [82aaad2]
+  - @inventorjs/core@3.9.2
+  - @inventorjs/plugin-app@1.7.2
+  - @inventorjs/plugin-plugin@1.4.2
+
 ## 1.5.3
 
 ### Patch Changes