+++ packages/plugin-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inventorjs/cli-plugin-plugin
 
+## 1.4.2
+
+### Patch Changes
+
+- 82aaad2: support context import plugin
+- Updated dependencies [82aaad2]
+  - @inventorjs/core@3.9.2
+
 ## 1.4.1
 
 ### Patch Changes